backend/requirements.txt
# ── Core Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Agent Framework ──
langgraph==0.2.60
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
google-generativeai==0.8.4

# ── Database ──
supabase==2.11.0
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10

# ── Redis ──
redis==5.2.1
aioredis==2.0.1

# ── Vector Database ──
pinecone-client==5.0.1

# ── External APIs ──
httpx==0.28.1
aiohttp==3.11.11
sendgrid==6.11.0
twilio==9.4.0
deepgram-sdk==3.8.0
elevenlabs==1.16.0

# ── Utilities ──
phonenumbers==8.13.52
python-dateutil==2.9.0
python-multipart==0.0.19
beautifulsoup4==4.12.3
Jinja2==3.1.5
email-validator==2.2.0

# ── WebSocket ──
python-socketio==5.12.1

# ── Testing ──
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# ── Security ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
